Keep Eigen factorization helpers in their owning implementation
